+++ src/Payment/Gateways/NullGateway.php
@@ -31,6 +31,11 @@ class NullGateway implements PaymentGateway
         return __('Offline');
     }
 
+    public static function svgIcon(): string
+    {
+        return '<svg viewBox="0 0 512 512"><path d="M256 512A256 256 0 1 0 256 0a256 256 0 1 0 0 512zm0-352a96 96 0 1 1 0 192 96 96 0 1 1 0-192z"/></svg>';
+    }
+
     public function createPaymentRequest(
         Payment $payment,
         Address $shippingAddress = null,