+++ src/Payment/Tests/Examples/UnorthodoxGateway.php
@@ -31,6 +31,11 @@ class UnorthodoxGateway implements PaymentGateway
         return 'Unorthodox Payments';
     }
 
+    public static function svgIcon(): string
+    {
+        return '';
+    }
+
     public function createPaymentRequest(
         Payment $payment,
         Address $shippingAddress = null,